fix: spawn npx process in a shell to fix ENOENT
When running the commands that are supposed to call `npx forge deploy/register/etc`, the following error is displayed on Windows machines: `spawn npx ENOENT`.

In order to fix it, spawning the npx process in a shell will fix the issue. See [this response](https://stackoverflow.com/a/59830014) on SO.